Add DDM support for special comment annotations. (#21)
This modifies the DDM parser so that it checks for possible tokens prior
to assuming that "//" or "/*" is a comment.

As part of this, a bug was discovered when a syntax definition token
contains only whitespace. This fixes the error.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.