Ion format simplifications; DDM elaboration changes to reduce panics. (#85)
This changes the Ion format to reduce size and make parsing more
permisive (so it is easier to generate in other languages).

This also implements some refactoring of elaboration code to reduce
panics.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.