Add #load_dialect command to DDM (#128)
This introduces a new DDM command elaborator, `#load_dialect`, that
loads a dialect from a file. It can be used to load example dialects in
source files.

As part of this, functionality from the Strata command line executable
was migrated into the DDM library code and simplified. The eventual goal
is to make it easy to work with dialects from other languages in Lean.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.