Add position info to DDM. (#112)
This is a large PR that adds the ability to attach annotations to DDM
subterms. This is used for source location information. This is
inherently a large change since many operators need an additional
argument and type level parameters are needed.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.

---------

Co-authored-by: Juneyoung Lee <lebjuney@amazon.com>
Co-authored-by: Juneyoung Lee <136006969+aqjune-aws@users.noreply.github.com>