Improve efficiency of creating a Strata program from source string. (#171)
This changes the DDM Lean environment extension so that we record where
the name the dialect is stored at in Lean and additionally associate a
map object with it.

This allows Strata programs to reuse the dialect map.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.